scsi_id: retry inquiry ioctl if host_byte is DID_TRANSPORT_DISRUPTED
The inquiry is issued to kernel via ioctl, kernelspace would set
this inquiry command's retry count to 0 which means the command
would not be retried in kernel space even if the LLDs returned
a status which need to be retried. So we should take the retry
in user space.

Signed-off-by: Wenchao Hao <haowenchao@huawei.com>
(cherry picked from commit f00f2837773fbfcd0f776506372ef39ff1da8e91)